[WP] reorganizing tests
